REFAC-02: Tweak I/O methods of game
The Game class was modified to overload its constructors, so that it can
 optionally allow specifying input and output streams, defaulting to no
input / no output. This makes it so that I/O streams only need to be
specified once during initialisation of the game, and only if they are
relevant; unit tests that do not make use of input or output may remain
unaltered.